Moved the handle tests
